Delete README.md
# Dentist Office Website

For the first six weeks at the Iron Yard, we each worked on redesigning a different local dentist's website.

We went through all the stages of the process:

*competitive market analysis
*background research
*illustrator style tiles
*illustrator wireframes
*web style tiles
*web wireframes

This is the culmination of all that work on my very first website. It's meant to be educational for the most part, which is why I am sure I will one day look back at this project and realize just how much my skills have improved with time and experience in the field.